Features, widget: don't use the RSS widget.
This widget requires a "url" parameter.

With CLI, this is transformed into a `--url=` parameter. This parameter is confused for CLI's global `--url=` parameter,
so we set the environment variable `WP_CLI_STRICT_ARGS_MODE=1`.

With CLI-over-SSH, CLI does not support setting that environment on the remote server.
For now, we'll change the widget and consider submitting a PR to the WP-CLI project to support this.