Add test for user creation error when using WPCLI
Add a new scenario in `user.feature` which creates a new user, logs in
as that user and then check the login name is displayed correctly on the
dashboard.

Create new step in DashboardContext to check the user name. Also create
new elements in AdminPage to expose the associated user name
information.

This test is currently failing owing to issue-126.



Tests: #126