Declare awareness trait contract
I noticed that my IDE couldn't resolve the call to getDriver() in
any of the AwarenessTraits. This is because getDriver() is
declared in RawWordPressContext which is not part of the
AwarenessTrait hierachy. I have therefore created
BaseAwarenessTrait which contains a single abstract method
getDriver(). This means if an AwarenessTrait is included in a
class which does not implement getDriver() there will be a
compliation time failure - rather than a messy runtime debug. It
also means that IDEs know what getDriver is and can make proper
use of the type hinting we've included.