Clear the database cache subsequent to restoration.
The wpphp driver wasn't clearing the WordPress database cache after it
had restored the database. This meant that users which had been
successfully removed from the database were still present in the db
cache. Wordpress will never login from a cached user - it always
interogrates the database, so the code to detect if the user was aready
present was seeing the cached user; but the login code wasn't. Hence the
failure.