Context, Dashboard: remove certain step definition matches for `iShouldBeOnThePage()`.
While we interact with the Dashboard via `PageObject`, to help distinguish between references to it, and us meaning the actual "page" or "screen" we see in a web browser, we should use "screen" to talk about UI screens as opposed to pages (unless it makes more sense to do so otherwise).

A few step matches have also been removed for brevity. We should take the opportunity to define how we want steps to be written to interact with WordHat, and not try to cover every possible user edge case.

See #152