Drivers: set `path` argument as optional.
To facilitate environments where the WordPress path is not known until runtime, and where CLI-over-SSH is used, the `path` argument is now optional.

I am not sure why `path` used to be checked with `realpath()`; it was implemented in the earliest stagest of WordHat's development, and it might not have had a reason that stands up to scrutiny at the current time.

See #185